requirements-security.txt
# Core Dependencies
flask>=2.0.1
flask-cors>=3.0.10
flask-limiter>=2.4.0
python-dotenv>=0.19.0

# Security Dependencies
pyOpenSSL>=20.0.1
cryptography>=35.0.0
requests>=2.26.0

# Network Monitoring
scapy>=2.4.5
netifaces>=0.11.0

# Google API
google-api-python-client>=2.27.0
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0

# Additional Security
passlib>=1.7.4
bcrypt>=3.2.0
PyJWT>=2.3.0

# Utilities
python-dateutil>=2.8.2
pytz>=2021.3

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1
